View and Submit an idea
Front end- View list of idea and click on submit idea for pop up to show up. Code is fully commented, but do message me if you have any queries.
Logo is missing & and I will add it later- if anyone wants to do that feel free to do so.